Add workaround for dupe signal issue on blueprint update
Issue:
https://mods.factorio.com/mod/cybersyn-combinator/discussion/67346043d8d01506c6ded0e5

Apparently in some circumstances when updating a blueprint with
a combinator in it, there will be additional sections added, sometimes
with duplicate signals.

There might be a better fix for this that involves reacting to
blueprint updates (if possible), for now we'll just improve the
sorting function to account for existing duplicate signals and just
overwrite them with the most recently processed one.