Improve display of special filter values
If a section somehow gets non-constant signals in it, they would show
up weirdly in the combinator making it confusing for the player.

Now we re-implement some base game UI to show them properly.
(Almost, we have to show the comparator in the top left because we
can't move the quality icon in the bottom left.)