Replace hack with new official API
There is currently a (probably) bug that makes it not possible to get
the count for a logistic group:
https://forums.factorio.com/viewtopic.php?t=129484

This is just a visual issue though.

Because this is an API added in 2.0.56, the version requirement for
the `base` dependency has been increased. (To 2.0.57 because 2.0.56
had an issue requiring it to be pulled, so the actual latest available
Factorio version that has the new API is 2.0.57.)